Fix #26
